@@ -2,113 +2,45 @@ using System.Diagnostics.CodeAnalysis;
 
 namespace ErrorOr;
 
-/// <summary>
-/// A discriminated union of errors or a value.
-/// </summary>
-/// <typeparam name="TValue">The type of the underlying <see cref="Value"/>.</typeparam>
 public readonly partial record struct ErrorOr<TValue> : IErrorOr<TValue>
 {
-    private readonly TValue? _value = default;
-    private readonly List<Error>? _errors = null;
-
-    /// <summary>
-    /// Prevents a default <see cref="ErrorOr"/> struct from being created.
-    /// </summary>
-    /// <exception cref="InvalidOperationException">Thrown when this method is called.</exception>
     public ErrorOr()
     {
         throw new InvalidOperationException("Default construction of ErrorOr<TValue> is invalid. Please use provided factory methods to instantiate.");
     }
 
-    private ErrorOr(Error error)
+    private ErrorOr(params IReadOnlyList<Error> errors)
     {
-        _errors = [error];
-    }
+        ArgumentNullException.ThrowIfNull(errors);
 
-    private ErrorOr(List<Error> errors)
-    {
-        if (errors is null)
-        {
-            throw new ArgumentNullException(nameof(errors));
-        }
-
-        if (errors is null || errors.Count == 0)
+        if (errors.Count == 0)
         {
             throw new ArgumentException("Cannot create an ErrorOr<TValue> from an empty collection of errors. Provide at least one error.", nameof(errors));
         }
 
-        _errors = errors;
+        Errors = errors;
     }
 
     private ErrorOr(TValue value)
     {
-        if (value is null)
-        {
-            throw new ArgumentNullException(nameof(value));
-        }
+        ArgumentNullException.ThrowIfNull(value);
 
-        _value = value;
+        Value = value;
+        Errors = [];
     }
 
-    /// <summary>
-    /// Gets a value indicating whether the state is error.
-    /// </summary>
-    [MemberNotNullWhen(true, nameof(_errors))]
-    [MemberNotNullWhen(true, nameof(Errors))]
+    public IReadOnlyList<Error> Errors { get; }
+
+    public Error? FirstError => IsError ? Errors[0] : null;
+
     [MemberNotNullWhen(false, nameof(Value))]
-    [MemberNotNullWhen(false, nameof(_value))]
-    public bool IsError => _errors is not null;
+    [MemberNotNullWhen(true, nameof(FirstError))]
+    public bool IsError => Errors.Count > 0;
 
-    /// <summary>
-    /// Gets the list of errors. If the state is not error, the list will contain a single error representing the state.
-    /// </summary>
-    /// <exception cref="InvalidOperationException">Thrown when no errors are present.</exception>
-    public List<Error> Errors => IsError ? _errors : throw new InvalidOperationException("The Errors property cannot be accessed when no errors have been recorded. Check IsError before accessing Errors.");
+    public TValue? Value { get; }
 
-    /// <summary>
-    /// Gets the list of errors. If the state is not error, the list will be empty.
-    /// </summary>
-    public List<Error> ErrorsOrEmptyList => IsError ? _errors : EmptyErrors.Instance;
-
-    /// <summary>
-    /// Gets the value.
-    /// </summary>
-    /// <exception cref="InvalidOperationException">Thrown when no value is present.</exception>
-    public TValue Value
+    public static ErrorOr<TValue> From(IReadOnlyList<Error> errors)
     {
-        get
-        {
-            if (IsError)
-            {
-                throw new InvalidOperationException("The Value property cannot be accessed when errors have been recorded. Check IsError before accessing Value.");
-            }
-
-            return _value;
-        }
-    }
-
-    /// <summary>
-    /// Gets the first error.
-    /// </summary>
-    /// <exception cref="InvalidOperationException">Thrown when no errors are present.</exception>
-    public Error FirstError
-    {
-        get
-        {
-            if (!IsError)
-            {
-                throw new InvalidOperationException("The FirstError property cannot be accessed when no errors have been recorded. Check IsError before accessing FirstError.");
-            }
-
-            return _errors[0];
-        }
-    }
-
-    /// <summary>
-    /// Creates an <see cref="ErrorOr{TValue}"/> from a list of errors.
-    /// </summary>
-    public static ErrorOr<TValue> From(List<Error> errors)
-    {
-        return errors;
+        return new(errors);
     }
 }
