@@ -12,7 +12,7 @@ public static partial class ErrorOrExtensions
     /// <param name="onValue">The action to execute if the state is a value.</param>
     /// <param name="onError">The action to execute if the state is an error.</param>
     /// <returns>The result of the executed function.</returns>
-    public static async Task Switch<TValue>(this Task<ErrorOr<TValue>> errorOr, Action<TValue> onValue, Action<List<Error>> onError)
+    public static async Task Switch<TValue>(this Task<ErrorOr<TValue>> errorOr, Action<TValue> onValue, Action<IReadOnlyList<Error>> onError)
     {
         var result = await errorOr.ConfigureAwait(false);
 
@@ -29,7 +29,7 @@ public static partial class ErrorOrExtensions
     /// <param name="onValue">The action to execute if the state is a value.</param>
     /// <param name="onError">The action to execute if the state is an error.</param>
     /// <returns>The result of the executed function.</returns>
-    public static async Task SwitchAsync<TValue>(this Task<ErrorOr<TValue>> errorOr, Func<TValue, Task> onValue, Func<List<Error>, Task> onError)
+    public static async Task SwitchAsync<TValue>(this Task<ErrorOr<TValue>> errorOr, Func<TValue, Task> onValue, Func<IReadOnlyList<Error>, Task> onError)
     {
         var result = await errorOr.ConfigureAwait(false);
 
