@@ -87,7 +87,7 @@ public class ElseTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Unexpected);
+        result.FirstError?.Type.Should().Be(ErrorType.Unexpected);
     }
 
     [Fact]
@@ -121,7 +121,7 @@ public class ElseTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Unexpected);
+        result.FirstError?.Type.Should().Be(ErrorType.Unexpected);
     }
 
     [Fact]
@@ -155,7 +155,7 @@ public class ElseTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Unexpected);
+        result.FirstError?.Type.Should().Be(ErrorType.Unexpected);
     }
 
     [Fact]
@@ -223,7 +223,7 @@ public class ElseTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Unexpected);
+        result.FirstError?.Type.Should().Be(ErrorType.Unexpected);
     }
 
     [Fact]
@@ -240,7 +240,7 @@ public class ElseTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Unexpected);
+        result.FirstError?.Type.Should().Be(ErrorType.Unexpected);
     }
 
     [Fact]
@@ -257,6 +257,6 @@ public class ElseTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Unexpected);
+        result.FirstError?.Type.Should().Be(ErrorType.Unexpected);
     }
 }
