@@ -17,7 +17,7 @@ public class FailIfTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Failure);
+        result.FirstError?.Type.Should().Be(ErrorType.Failure);
     }
 
     [Fact]
@@ -33,7 +33,7 @@ public class FailIfTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Failure);
+        result.FirstError?.Type.Should().Be(ErrorType.Failure);
     }
 
     [Fact]
@@ -63,7 +63,7 @@ public class FailIfTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.NotFound);
+        result.FirstError?.Type.Should().Be(ErrorType.NotFound);
     }
 
     [Fact]
@@ -78,8 +78,8 @@ public class FailIfTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Failure);
-        result.FirstError.Description.Should().Be("5 is greater than 3");
+        result.FirstError?.Type.Should().Be(ErrorType.Failure);
+        result.FirstError?.Description.Should().Be("5 is greater than 3");
     }
 
     [Fact]
@@ -95,8 +95,8 @@ public class FailIfTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.Failure);
-        result.FirstError.Description.Should().Be("5 is greater than 3");
+        result.FirstError?.Type.Should().Be(ErrorType.Failure);
+        result.FirstError?.Description.Should().Be("5 is greater than 3");
     }
 
     [Fact]
@@ -126,6 +126,6 @@ public class FailIfTests
 
         // Assert
         result.IsError.Should().BeTrue();
-        result.FirstError.Type.Should().Be(ErrorType.NotFound);
+        result.FirstError?.Type.Should().Be(ErrorType.NotFound);
     }
 }
